feature: sync sym links closes #51
